Preparando para merge

+++ Pprueba/Src/Game.h
@@ -21,7 +21,7 @@ public:
 	bool stop();
 
 	inline float getDeltaTime() const {
-		return deltaTime;
+		return (float)deltaTime;
 	};
 
 private:
